Renamed CrimeConsumer to CrimeClient for consistent naming structure. Also merged Scraper into CrimeClient.
