+++ tests/HDS/CrimeClientTests.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * Defines CrimeConsumer unit tests.
+ */
+
+
+/**
+ * Unit Tests for CrimConsumer
+ */
+final class CrimeConsumerTests{
+
+     public function consume_BasicUse()
+     {
+     }
+ }
