--- tests/HDS/CrimeConsumerTests.php
@@ -1,41 +1,0 @@
-<?php
-/**
- * Defines CrimeConsumer unit tests.
- */
-
-
-/**
- * Unit Tests for CrimConsumer
- */
-final class CrimeConsumerTests{
-
-     public function insertVABeachRecord()
-     {
-
-     }
-
-     public function insertNorfolkRecord()
-     {
-
-     }
-
-     public function insertNpnRecord()
-     {
-
-     }
-
-     public function insertChesapeake()
-     {
-
-     }
-
-     public function insertSuffolkRecord()
-     {
-
-     }
-
-     public function insertPortsmouthRecord()
-     {
-
-     }
- }
